Replaced gtags-parser(1) with global(1) with the -f command.
Htags(1) should not use gtags-parser(1) directly; it should use
global(1) with the -f option instead. Because gtags-parser is
part of the implementation of gtags(1) and global(1), and htags
is only one of application programs which use global(1).
If htags depends on gtags-parser, it will become difficult to
change the implementation of gtags and global.